refactor: flatten macro runtime
