+++ src/choiceExecutor.ts
@@ -7,7 +7,7 @@ import type IMacroChoice from "./types/choices/IMacroChoice";
 import { TemplateChoiceEngine } from "./engine/TemplateChoiceEngine";
 import { CaptureChoiceEngine } from "./engine/CaptureChoiceEngine";
 import { MacroChoiceEngine } from "./engine/MacroChoiceEngine";
-import { SingleInlineScriptEngine } from "./engine/SingleInlineScriptEngine";
+import { InlineJavaScriptEvaluator } from "./engine/InlineJavaScriptEvaluator";
 import { SingleMacroEngine } from "./engine/SingleMacroEngine";
 import type { IChoiceExecutor } from "./IChoiceExecutor";
 import type { FormatterEvaluatorContext } from "./formatters/formatterEvaluators";
@@ -56,7 +56,7 @@ export class ChoiceExecutor implements IChoiceExecutor {
 		code: string,
 		context: FormatterEvaluatorContext,
 	): Promise<unknown> {
-		const executor = new SingleInlineScriptEngine(
+		const executor = new InlineJavaScriptEvaluator(
 			this.app,
 			this.plugin,
 			this,
