--- src/engine/SingleInlineScriptEngine.ts
@@ -1,30 +1,0 @@
-import type { App } from "obsidian";
-import type QuickAdd from "../main";
-import type { IChoiceExecutor } from "../IChoiceExecutor";
-import { MacroChoiceEngine } from "./MacroChoiceEngine";
-
-type AsyncFunctionConstructor = new (code: string) => () => Promise<unknown>;
-
-export class SingleInlineScriptEngine extends MacroChoiceEngine {
-	constructor(
-		app: App,
-		plugin: QuickAdd,
-		choiceExecutor: IChoiceExecutor,
-		variables: Map<string, unknown>
-	) {
-		//@ts-ignore
-		super(app, plugin, null, choiceExecutor, variables);
-	}
-
-	 
-	public async runAndGetOutput(code: string): Promise<unknown> {
-		const AsyncFunction = Object.getPrototypeOf(
-			async function () {}
-		).constructor as AsyncFunctionConstructor;
-		 
-		const userCode = new AsyncFunction(code);
-
-		 
-		return await userCode.bind(this.params, this).call();
-	}
-}
